V3.1
Added Editing Features
Improved deletion

+++ delete.php
@@ -1,34 +1,41 @@
 <?php
 session_start();
 ob_start();
-require_once('./include/db_connect.php');
+require_once './include/db_connect.php';
+require_once('helper.php');
 
-print_r( $_GET );
+//print_r( $_GET );
+
+if ( getAdminStatus()=='you are an admin' ) {
 
 
-if ( $_GET['id'] ) { //echo "we have an id";
 
-	$id = $_GET['id'];
+	if ( $_GET['id'] ) { //echo "we have an id";
 
-	// build the query
-	$query = "DELETE FROM users WHERE id = " . $id;
+		$id = $_GET['id'];
 
-	$result = mysql_query( $query ) or die( 'Could not delete the user.' );
-	if ( $result ) {
-		header( "Location: http://".$_SERVER['HTTP_HOST'].dirname( $_SERVER['PHP_SELF'] )."/"."admin.php" );
+		// build the query
+		$query = "DELETE FROM users WHERE id = " . $id;
+
+		$result = mysql_query( $query ) or die( 'Could not delete the user.' );
+		if ( $result ) {
+			header( "Location: http://".$_SERVER['HTTP_HOST'].dirname( $_SERVER['PHP_SELF'] )."/"."admin.php" );
 			exit();
-	}else {
-		echo "Something went horribly wrong when attempting to delete ID " . $id ."<br/>";
+		}else {
+			echo "Something went horribly wrong when attempting to delete ID " . $id ."<br/>";
+		}
+
+	} else {   // echo "we don't have an id";
+
+
+
 	}
 
-} else {   // echo "we don't have an id";
 
+} else {
+	//header( "Location: http://".$_SERVER['HTTP_HOST'].dirname( $_SERVER['PHP_SELF'] )."/include/fu.html" );
 
+	include('./include/fu.html');
 
 }
-
-
-
-
-
 ?>
